Conversion improvements. Unit refactor.
Unit is now a sealed class with member classes: Base and Converted
Converted now uses a new bidirectional conversion class to move between itself and base
units are now properties of the unit systems, not objects. there was no advantage to them being objects, and this way they are now added to the unit system immediately as opposed to when lazily loaded.